% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descritivas.R
\name{calcular_cv}
\alias{calcular_cv}
\title{Calcular o Coeficiente de Variação (CV)}
\usage{
calcular_cv(x, na.rm = FALSE)
}
\arguments{
\item{x}{Um vetor numérico.}

\item{na.rm}{Lógico. Se TRUE, remove valores NA antes do cálculo.}
}
\value{
Um número representando a porcentagem do coeficiente de variação.
}
\description{
Esta função calcula a variabilidade relativa dos dados.
A fórmula utilizada é:
\deqn{CV = \frac{s}{\bar{x}} \times 100}
}
\examples{
dados <- c(10, 20, 30, 40)
calcular_cv(dados)
}
