% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descritivas.R
\name{construir_hist}
\alias{construir_hist}
\title{Constrói Histograma Simples}
\usage{
construir_hist(x, titulo = "Histograma")
}
\arguments{
\item{x}{Um vetor numérico contendo os dados.}

\item{titulo}{Uma string para o título do gráfico.}
}
\value{
Um objeto ggplot que pode ser exibido.
}
\description{
Esta função utiliza o ggplot2 para criar um histograma simples
para um vetor numérico.
}
\examples{
dados <- rnorm(100, mean = 50, sd = 10)
construir_hist(dados, "Distribuição de Dados de Exemplo")
}
