% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saude.R
\name{imc}
\alias{imc}
\title{Calcula o Índice de Massa Corporal (IMC)}
\usage{
imc(peso, altura)
}
\arguments{
\item{peso}{Um valor numérico representando o peso em quilogramas (kg).}

\item{altura}{Um valor numérico representando a altura em metros (m).}
}
\value{
Um valor numérico representando o Índice de Massa Corporal (IMC).
}
\description{
Esta função calcula o IMC usando a fórmula peso / altura^2.
}
\examples{
# IMC de uma pessoa com 70kg e 1.75m
imc(peso = 70, altura = 1.75)
}
