% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descritivas.R
\name{resumo_stats}
\alias{resumo_stats}
\title{Resumo Estatístico Personalizado}
\usage{
resumo_stats(x)
}
\arguments{
\item{x}{Um vetor numérico.}
}
\value{
Um data frame com as estatísticas descritivas.
}
\description{
Retorna média, mediana, desvio padrão e erro padrão.
}
